requirements.txt
# 動物園環境教育 Line Bot - 套件清單
# Python 3.9+

# ============================================================
# 後端框架
# ============================================================
Flask==3.0.0
gunicorn==21.2.0

# ============================================================
# Line Bot
# ============================================================
line-bot-sdk==3.5.0

# ============================================================
# OpenAI
# ============================================================
openai==1.3.0

# ============================================================
# 深度學習
# ============================================================
torch==2.1.0
transformers==4.35.0

# ============================================================
# 資料處理
# ============================================================
pandas==2.1.0
numpy==1.24.0

# ============================================================
# 資料庫
# ============================================================
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9

# ============================================================
# 排程
# ============================================================
APScheduler==3.10.4

# ============================================================
# HTTP 請求
# ============================================================
requests==2.31.0

# ============================================================
# 環境變數
# ============================================================
python-dotenv==1.0.0

# ============================================================
# 資料視覺化
# ============================================================
matplotlib==3.8.0
seaborn==0.13.0

# ============================================================
# 機器學習工具
# ============================================================
scikit-learn==1.3.0

# ============================================================
# 其他
# ============================================================
jieba==0.42.1
python-dateutil==2.8.2
